blog/requirements.txt
attrs
jupyter
jinja2
tomlkit